requirements.txt
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
xlrd == 2.0.1

pip~=25.1
openpyxl~=3.1.5
lxml~=5.4.0
matplotlib~=3.10.1
scipy~=1.15.2
pillow~=11.2.1
et_xmlfile~=2.0.0
xlwt~=1.3.0
mysqlclient~=2.2.7
mysql~=0.0.3
mysql-connector~=2.2.9
mysql-connector-python~=9.2.0